requirements.txt
requests
pyyaml
python-telegram-bot==20.3
